+++ examples/rails/app/commands/users_commands/authentication/login_command.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module UsersCommands
+
+  module Authentication
+
+    # LoginCommand
+    class LoginCommand < ApplicationCommand
+
+      validates :name, type: :string
+
+    end
+
+  end
+
+end